refactor(solver): split mapped declaration surface queries
